*** empty log message ***
git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk@2777 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ M2/Macaulay2/e/matrix.cpp
@@ -722,7 +722,7 @@ Matrix *Matrix::random(const Ring *R,
   if (fraction_non_zero != 1.0)
     {
       doing_fraction = true;
-      int f = fraction_non_zero * 10000;
+      int f = static_cast<int>(fraction_non_zero * 10000);
       threshold = 10000 - f;
       printf("threshold is %d\n", threshold);
     }
